requirements.txt
asgiref==3.5.2
attrs==21.4.0
cached-property==1.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==4.0.5
django-ckeditor==6.4.2
django-import-export==2.8.0
django-js-asset==2.0.0
et-xmlfile==1.1.0
filelock==3.7.1
gunicorn==20.1.0
idna==3.3
MarkupPy==1.14
mysqlclient==2.1.0
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
platformdirs==2.5.2
psycopg2==2.9.3
pyparsing==3.0.9
pythonfinder==1.2.10
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
